refactor: Add initiateOAuth function for OAuth authentication
The `initiateOAuth` function has been added to the `helpers.ts` file. This function allows users to initiate OAuth authentication with either Google or GitHub providers. It redirects the user to the appropriate callback URL based on the selected provider. This addition improves the authentication flow and enhances the user experience.